[256936] Support for Offline Mode https://bugs.eclipse.org/bugs/show_bug.cgi?id=256936

+++ plugins/org.eclipse.emf.cdo/src/org/eclipse/emf/cdo/session/CDOSession.java
@@ -197,7 +197,7 @@ public interface CDOSession extends CDOCommonSession, IContainer<CDOView>
   /**
    * Refreshes the object caches of all (non-historical) {@link CDOView views}.
    */
-  public int refresh();
+  public long refresh();
 
   /**
    * Returns the time stamp of the last commit operation. May not be accurate if
